Add All Transit
Hello, I made [a visualization](https://all-transit.com) of transit using Transitland data, and it looks like it would be a good addition to the list. I added it right below `TransitFlow`, because that's the most similar project, but I could move it to the bottom of the list if you preferred.